@@ -22,13 +22,13 @@
 	"minimum-stability": "dev",
 	"prefer-stable": true,
 	"require": {
-		"php": ">=7.2",
-		"icanboogie/icanboogie": "^5.0",
+		"php": ">=8.1",
+		"icanboogie/icanboogie": "^6.0",
 		"icanboogie/render": "^0.7|^0.8"
 	},
 	"require-dev": {
 		"phpstan/phpstan": "^1.5",
-		"phpunit/phpunit": "^8.5"
+		"phpunit/phpunit": "^9.5"
 	},
 	"autoload": {
 		"psr-4": {
